Party time.

+++ materials_lighting/shaders/fragment.glsl
@@ -9,7 +9,7 @@ struct Material {
     vec3 ambient;
     vec3 diffuse;
     vec3 specular;
-    float shininess;
+    int shininess;
 }; 
   
 uniform Material material;
